+++ src/usecases/catalogo.rs
@@ -2,6 +2,7 @@ use std::sync::Arc;
 
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
+use rust_decimal::Decimal;
 
 use crate::{models::{Adicional, Produto, Usuario}, services::CatalogoService};
 
@@ -9,7 +10,7 @@ use crate::{models::{Adicional, Produto, Usuario}, services::CatalogoService};
 pub struct AtualizarProdutoRequest {
     nome: String,
     descricao: Option<String>,
-    preco: f64,
+    preco: Decimal,
     categoria_uuid: Uuid,
     tempo_preparo_min: Option<i32>,
 }
@@ -20,7 +21,7 @@ pub struct CreateProdutoRequest {
     pub categoria_uuid: Uuid,
     pub nome: String,
     pub descricao: Option<String>,
-    pub preco: f64,
+    pub preco: Decimal,
     pub imagem_url: Option<String>,
     pub disponivel: bool,
     pub tempo_preparo_min: Option<i32>,
